Create port scanner
This example takes the hostname and port number as command-line arguments and uses the socket API to check if the specified port on the target host is open or closed. This is done by creating a socket and trying to connect to the target host on the specified port. If the connection is successful, the port is considered open, otherwise it is considered closed.